Map 2 Progress. (#21)
* More buildings + pathnodes

Decided to play around with pathnodes and added them in. Not really sure how they are supposed to work, but they are there. Also done some more buildings (Jack In Communcations and more rural house)

* Map 2 progress

It's still wip, a lot of stuff was for testing. But the outline for the border mountains and the church is ready.