Overhauls?
So the changes are that I smoothed out the dog code, and made it move smoothly. Still has speed issues like cats but I think that can be resolved. As for the maps - I cleared world1 off navigation actors and I added in to both maps the Hub cross line. That's all